Misc. Enhancements.
Upload files that can be removed once uploaded first in order to
help save space locally.

Check to make sure two renames or uploads won't end up with the same
file name on the remote.  To make this operation efficient add a unique
index to for the remote_path field in the fileinfo table.  Increment
schema version.

Other misc. changes.