Create a basis preprocessor that scans for SQL macros.
The preprocessor works by scanning syncfs.cpp for the "SQL" keyword.
It then takes the SQL string and stories it in an array.  The macro is
then replaced by code that looks up the query in the array.  The
preprocessor is careful to preserve both line and column information.

Local variables can be used directly in queries by using a '$'.  This
works by converting the variable to a bound parameter.

The main use for the preprocessor is to make it easy to prepare all
the queries on program startup to quickly catch any obvious problems.
The other reason is to make working with SQL easier.