Enhance the SQL preprocessor.
In addition to recognizing the SQL macro, the EXEC and SELECT macros
are recognized.  EXEC will execute a non-select query (rather than just
preparing it) and SELECT will return a SqlResult (but it won't be
executed until the step() method is called.

The preprocessor will also parse "select" queries for a list of
columns.  If the columns are named and it is able to figure out the
C++ type to represent it than an enhanced SqlResult is created that
supports getting the columns via a C structure.  Types for columns can
be specified via the the special '`' syntax or by adding the column
name to a lookup table.

The entire schema is now also stored in a single file that is
converted to a string constant by the preprocessor.