More Enhancements to the SQL preprocessor.
(1) Rather than using a structure to return the results of an select query
with just one column, use the column type directly.  This allows
unnamed columns to also be accessed directly as long as the type can
be determined.

(2) Create an iterator for iterating over the enhanced SqlResult.
The main benefit of this is to support directly iterating over the
querie result using a C++11 "for" loops.

(3) Support '.','->', and '::' as part of the name for when inlining
variables into queries via the '$' syntax.

Also enhance syncsfs.cpp to take full advantage of the SQL
preprocessor features.